Adds Preview component
